refactor(oscript): перенести классы поддержки extends в подпакет extends_
OScriptExtends, TypeRelations и TypeRelationIndex — поддержка библиотеки
наследования extends; выделены в подпакет types.oscript.extends_
(extends — зарезервированное слово Java).
