+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/oscript/extends_/TypeRelationIndex.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.types.oscript;
+package com.github._1c_syntax.bsl.languageserver.types.oscript.extends_;
 
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
 import com.github._1c_syntax.bsl.languageserver.context.FileType;
@@ -37,6 +37,8 @@ import java.util.LinkedHashSet;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
+import com.github._1c_syntax.bsl.languageserver.types.oscript.OScriptLibraryIndex;
+import com.github._1c_syntax.bsl.languageserver.types.oscript.AbstractOScriptLazyIndex;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
