+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/types/oscript/extends_/TypeRelationsTest.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.types.oscript;
+package com.github._1c_syntax.bsl.languageserver.types.oscript.extends_;
 
 import com.github._1c_syntax.bsl.languageserver.context.AbstractServerContextAwareTest;
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
@@ -34,6 +34,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.net.URI;
 import java.nio.file.Path;
 
+import com.github._1c_syntax.bsl.languageserver.types.oscript.OScriptLibraryIndex;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
